2 major bugs and code clean up
There was 2 major bugs and a code clean up.

1. The first bug was that 0x37 didn't exist and 0x36 was in the zeroToNine byte array twice
2. The second bug was programming logic that made the program only read two digits before decimal and one after and the current version of BTD6 is 19.2 so I assumed it was right but after review corrections were needed. If you want to know the exact bug it was because on line 58 on the current release I had it to make continueSearchingBefore equal false so it was just a coincidence. Bug is fixed so no need to worry ;)

The code clean up was that last version had separate methods for FromDecimal for forward and backwards so instead I did a from decimal and use an argument to use the same method.
I just noticed there was some duplicated code and removed it. (This is difference between last update and this update ik I said I was done lol)

Less than 100 lines of code (97) last one was 107 and had bugs. This will most likely be last update as it works perfectly and is completely future proof.
The reason it is future proof is because it avoids the header and looks for a number after decimal (0x2E) and no other things do this except the header.
And even if they did add something will a number after a decimal this returns on the first detection. But if they change it in the future this will get updated again!